feat: Add timeline components for order management
- Introduced TimeIndicator.vue to display the current time on the timeline.
- Created TimelineOrderCard.vue for rendering individual orders with details.
- Added TimelineSlots.vue to visualize time slots on the timeline.
- Implemented utility functions in timelineUtils.ts for time formatting, conversion, and slot calculation.
- Removed date selection functionality from OrdinazioniView.vue to streamline the interface.
